refactor(auth): move auth related functions to auth.js
